+++ build.dep.network.gradle
@@ -3,6 +3,7 @@ apply {
 }
 
 dependencies {
+    implementation project(':core:utils')
     implementation libs.retrofit
     implementation(libs.kotlinx.serialization.json)
     implementation libs.retrofit2.kotlinx.serialization.converter
