+++ build.module.feature-and-app.gradle
@@ -30,6 +30,7 @@ android {
 }
 
 dependencies {
+    implementation project(':core:utils')
     implementation(libs.napier)
     implementation(libs.activityCompose)
     implementation(libs.hiltNavigationCompose)
