+++ core/network/src/main/kotlin/nl/q42/template/core/network/interceptor/UserAgentHeaderInterceptor.kt
@@ -1,8 +1,8 @@
 package nl.q42.template.core.network.interceptor
 
 import android.os.Build
-import nl.q42.template.core.network.di.ConfigAppVersionCode
-import nl.q42.template.core.network.di.ConfigAppVersionName
+import nl.q42.template.core.utils.di.ConfigAppVersionCode
+import nl.q42.template.core.utils.di.ConfigAppVersionName
 import okhttp3.Interceptor
 import okhttp3.Response
 import javax.inject.Inject
